improve: restructure gdocs files
