+++ db/model/Gdoc/gdocUtils.ts
@@ -1,0 +1,248 @@
+import {
+    Span,
+    EnrichedBlockText,
+    OwidEnrichedArticleBlock,
+    SpanLink,
+    SpanBold,
+    SpanItalic,
+    SpanFallback,
+    SpanQuote,
+    SpanSuperscript,
+    SpanSubscript,
+    SpanUnderline,
+    SpanRef,
+    EnrichedBlockSimpleText,
+    SpanSimpleText,
+} from "@ourworldindata/utils"
+import { match } from "ts-pattern"
+import { partition, compact } from "lodash"
+import * as cheerio from "cheerio"
+
+export function appendDotEndIfMultiline(line: string): string {
+    if (line.includes("\n")) return line + "\n.end"
+    return line
+}
+
+export function consolidateSpans(
+    blocks: OwidEnrichedArticleBlock[]
+): OwidEnrichedArticleBlock[] {
+    const newBlocks: OwidEnrichedArticleBlock[] = []
+    let currentBlock: EnrichedBlockText | undefined = undefined
+    for (const block of blocks) {
+        if (block.type === "text")
+            if (currentBlock === undefined) currentBlock = block
+            else
+                currentBlock = {
+                    type: "text",
+                    value: [...currentBlock.value, ...block.value],
+                    parseErrors: [],
+                }
+        else {
+            if (currentBlock !== undefined) {
+                newBlocks.push(currentBlock)
+                currentBlock = undefined
+                newBlocks.push(block)
+            }
+        }
+    }
+    return newBlocks
+}
+
+export function spanToHtmlString(s: Span): string {
+    return match(s)
+        .with({ spanType: "span-simple-text" }, (span) => span.text)
+        .with(
+            { spanType: "span-link" },
+            (span) =>
+                `<a href="${span.url}">${spansToHtmlString(span.children)}</a>`
+        )
+        .with(
+            { spanType: "span-ref" },
+            (span) =>
+                `<a href="${span.url}" class="ref">${spansToHtmlString(
+                    span.children
+                )}</a>`
+        )
+        .with({ spanType: "span-newline" }, () => "</br>")
+        .with(
+            { spanType: "span-italic" },
+            (span) => `<i>${spansToHtmlString(span.children)}</i>`
+        )
+        .with(
+            { spanType: "span-bold" },
+            (span) => `<b>${spansToHtmlString(span.children)}</b>`
+        )
+        .with(
+            { spanType: "span-underline" },
+            (span) => `<u>${spansToHtmlString(span.children)}</u>`
+        )
+        .with(
+            { spanType: "span-subscript" },
+            (span) => `<sub>${spansToHtmlString(span.children)}</sub>`
+        )
+        .with(
+            { spanType: "span-superscript" },
+            (span) => `<sup>${spansToHtmlString(span.children)}</sup>`
+        )
+        .with(
+            { spanType: "span-quote" },
+            (span) => `<q>${spansToHtmlString(span.children)}</q>`
+        )
+        .with(
+            { spanType: "span-fallback" },
+            (span) => `<span>${spansToHtmlString(span.children)}</span>`
+        )
+        .exhaustive()
+}
+
+export function spansToHtmlString(spans: Span[]): string {
+    if (spans.length === 0) return ""
+    else {
+        const result = spans.map(spanToHtmlString).join("")
+        return result
+    }
+}
+
+function spanFallback(node: CheerioElement): SpanFallback {
+    return {
+        spanType: "span-fallback",
+        children: compact(node.children?.map(cheerioToSpan)) ?? [],
+    }
+}
+
+export function htmlToEnrichedTextBlock(html: string): EnrichedBlockText {
+    return {
+        type: "text",
+        value: htmlToSpans(html),
+        parseErrors: [],
+    }
+}
+
+export function htmlToSimpleTextBlock(html: string): EnrichedBlockSimpleText {
+    const spans = htmlToSpans(html)
+    const [simpleTextSpans, otherSpans] = partition(
+        spans,
+        (s) => s.spanType === "span-simple-text"
+    )
+    const simpleText: SpanSimpleText = {
+        spanType: "span-simple-text",
+        text: simpleTextSpans.map((s) => (s as SpanSimpleText).text).join(" "),
+    }
+    const parseErrors =
+        otherSpans.length > 0
+            ? [
+                  {
+                      message:
+                          "Formatted text fragments found in simple text block",
+                  },
+              ]
+            : []
+    return {
+        type: "simple-text",
+        value: simpleText,
+        parseErrors: parseErrors,
+    }
+}
+
+export function htmlToSpans(html: string): Span[] {
+    const $ = cheerio.load(html)
+    const nodes = $("body").contents().toArray()
+    return compact(nodes.map(cheerioToSpan)) ?? []
+}
+
+// Sometimes Google automatically linkifies a URL.
+// We always want the plaintext, not '<a href="www.ourworldindata.org">www.ourworldindata.org</a>'
+export function extractPlaintextUrl(html: string = ""): string {
+    if (html.trim().startsWith("http")) return html.trim()
+    const $ = cheerio.load(html)
+    return $("a").text()
+}
+
+export function cheerioToSpan(node: CheerioElement): Span | undefined {
+    if (node.type === "text")
+        // The regex replace takes care of the ArchieML escaping of :
+        return {
+            spanType: "span-simple-text",
+            text: node.data?.replace(/\\:/g, ":") ?? "",
+        }
+    else if (node.type === "tag") {
+        return match(node.tagName)
+            .with("a", (): SpanLink | SpanRef => {
+                const url = node.attribs.href
+                const className = node.attribs.class
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                if (className === "ref") {
+                    return { spanType: "span-ref", children, url }
+                }
+                return { spanType: "span-link", children, url }
+            })
+            .with("b", (): SpanBold => {
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                return { spanType: "span-bold", children }
+            })
+            .with("i", (): SpanItalic => {
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                return { spanType: "span-italic", children }
+            })
+            .with("br", (): Span => ({ spanType: "span-newline" }))
+            .with("cite", () => spanFallback(node))
+            .with("code", () => spanFallback(node)) // TODO: should get a style
+            .with(
+                "em",
+                (): SpanItalic => ({
+                    spanType: "span-italic",
+                    children: compact(node.children?.map(cheerioToSpan)) ?? [],
+                })
+            )
+            .with(
+                "q",
+                (): SpanQuote => ({
+                    spanType: "span-quote",
+                    children: compact(node.children?.map(cheerioToSpan)) ?? [],
+                })
+            )
+            .with("small", () => spanFallback(node))
+            .with("span", () => spanFallback(node))
+            .with("strong", (): SpanBold => {
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                return { spanType: "span-bold", children }
+            })
+            .with("sup", (): SpanSuperscript => {
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                return { spanType: "span-superscript", children }
+            })
+            .with("sub", (): SpanSubscript => {
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                return { spanType: "span-subscript", children }
+            })
+            .with("u", (): SpanUnderline => {
+                const children =
+                    compact(node.children?.map(cheerioToSpan)) ?? []
+                return { spanType: "span-underline", children }
+            })
+            .with("wbr", () => spanFallback(node))
+            .otherwise(() => {
+                return undefined
+            })
+    }
+    return undefined
+}
+
+export const getTitleSupertitleFromHeadingText = (
+    headingText: string
+): [string, string | undefined] => {
+    const VERTICAL_TAB_CHAR = "\u000b"
+    const [beforeSeparator, afterSeparator] =
+        headingText.split(VERTICAL_TAB_CHAR)
+
+    return [
+        afterSeparator || beforeSeparator,
+        afterSeparator ? beforeSeparator : undefined,
+    ]
+}
