+++ db/model/Gdoc/rawToEnriched.ts
@@ -60,10 +60,10 @@ import {
     htmlToEnrichedTextBlock,
     htmlToSimpleTextBlock,
     htmlToSpans,
+    getTitleSupertitleFromHeadingText,
 } from "./gdocUtils.js"
 import { match } from "ts-pattern"
 import { parseInt } from "lodash"
-import { getTitleSupertitleFromHeadingText } from "./gdocToArchieml.js"
 
 export function parseRawBlocksToEnrichedBlocks(
     block: OwidRawArticleBlock
